feat: migrate art command
